requirements.txt
numpy
opencv-python
moviepy
openai-whisper
transformers
torch
scikit-learn
matplotlib
pytest